pyproject.toml
[project]
name = "langcore-api"
version = "0.1.0"
description = "Queue-based document extraction API powered by FastAPI, Celery, and LangCore"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.129.0",
    "uvicorn[standard]>=0.35.0",
    "pydantic-settings>=2.13.0",
    "celery[redis]>=5.6.2",
    "redis>=5.0.3,<6.5",
    "httpx>=0.28.1",
    "flower>=2.0.1",
    "prometheus-client>=0.22.0",
    "prometheus-fastapi-instrumentator>=7.0.2",
    "tenacity>=9.1.2",
    "structlog>=25.1.0",
    "langcore",
    "langcore-audit",
    "langcore-guardrails",
    "langcore-litellm",
    "langcore-dspy",
    "langcore-rag",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=6.0",
    "ruff>=0.13.0",
]

# Ruff config lives in .ruff.toml (single source of truth)

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = [
    "--strict-markers",
    "--tb=short",
]
